Refactored for packaging
