--- agents/__init__.py
@@ -1,2 +1,0 @@
-from agents.basic.agent import BasicAgent
-from agents.agent import Agent, Memory
