--- envs/single_agents_envs/__init__.py
@@ -1,2 +1,0 @@
-from envs.single_agents_envs.random_crosses_and_noughts import RdCrossesAndNoughtsEnv
-from envs.single_agents_envs.random_nim import RdNimEnv
