--- envs/single_agents_envs/random_nim/__init__.py
@@ -1,1 +1,0 @@
-from envs.single_agents_envs.random_nim.game import RdNimEnv
