+++ learnrl/agents/__init__.py
@@ -1,0 +1,1 @@
+from learnrl.agents.table.agent import TableAgent
