+++ learnrl/agents/table/control.py
@@ -1,7 +1,6 @@
 """
 Control methodes to improve the policy based on value fonctions
 """
-
 import numpy as np
