+++ learnrl/environments/multi_agents/crosses_and_noughts/images/black_background.png
