+++ learnrl/environments/multi_agents/crosses_and_noughts/images/cross.png
