+++ learnrl/environments/multi_agents/crosses_and_noughts/images/empty_grid.png
