+++ learnrl/environments/single_agents/random_crosses_and_noughts/images/black_background.png
