+++ learnrl/environments/single_agents/random_crosses_and_noughts/images/empty_grid.png
