+++ learnrl/environments/single_agents/random_crosses_and_noughts/images/nought.png
