+++ learnrl/environments/single_agents/random_nim/__init__.py
@@ -1,0 +1,1 @@
+from learnrl.environments.single_agents.random_nim.game import RdNimEnv
