+++ tests/tests_agents/tests_basic_agent/test_evaluation.py
@@ -1,8 +1,8 @@
 import pytest
 import numpy as np
 
-from agents.basic.evaluation import Evaluation, MonteCarlo, TemporalDifference
-from agents.agent import Memory
+from learnrl.agents.table.evaluation import Evaluation, MonteCarlo, TemporalDifference
+from learnrl.core import Memory
 
 # Test basic evaluation features
 def test_name():
